Week7/requirements.txt
fire==0.5.0
pyyaml==6.0.1
confluent-kafka==2.3.0
